Player's stat setting 2-2
- Editting related scripts.
- Fixed the bug in stats' bar.

+++ RPGPractice_2021/Assets/Game/Scripts/00Managers/01UI/Stats/HpBar.cs
@@ -27,6 +27,6 @@ public class HpBar : MonoBehaviour
 
     public void SetSliderValue(int currentHp, int maxHp)
     {
-        slider.value = (currentHp/maxHp);
+        slider.value = ((float)currentHp/(float)maxHp);
     }
 }
